AggiuntaCarrello

+++ src/main/java/it/unisa/db/ProductDaoDataSource.java
@@ -34,7 +34,7 @@ public class ProductDaoDataSource implements IProductDao {
 			preparedStatement = connection.prepareStatement(insertSQL);
 			preparedStatement.setString(1, product.getName());
 			preparedStatement.setString(2, product.getDescription());
-			preparedStatement.setInt(3, product.getPrice());
+			preparedStatement.setDouble(3, product.getPrice());
 			preparedStatement.setInt(4, product.getQuantity());
 			preparedStatement.setString(5, product.getImage());
 
@@ -71,7 +71,7 @@ public class ProductDaoDataSource implements IProductDao {
 				bean.setCode(rs.getInt("CODE"));
 				bean.setName(rs.getString("NAME"));
 				bean.setDescription(rs.getString("DESCRIPTION"));
-				bean.setPrice(rs.getInt("PRICE"));
+				bean.setPrice(rs.getDouble("PRICE"));
 				bean.setQuantity(rs.getInt("QUANTITY"));
 				bean.setImage(rs.getString("IMAGE"));
 			}
@@ -120,7 +120,7 @@ public class ProductDaoDataSource implements IProductDao {
 	    try (Connection con = ds.getConnection(); PreparedStatement ps = con.prepareStatement(sql)) {
 	        ps.setString(1, product.getName());
 	        ps.setString(2, product.getDescription());
-	        ps.setInt(3, product.getPrice());
+	        ps.setDouble(3, product.getPrice());
 	        ps.setInt(4, product.getQuantity());
 	        ps.setString(5, product.getImage());
 	        ps.setInt(6, product.getCode());
@@ -150,7 +150,7 @@ public class ProductDaoDataSource implements IProductDao {
 				bean.setCode(rs.getInt("CODE"));
 				bean.setName(rs.getString("NAME"));
 				bean.setDescription(rs.getString("DESCRIPTION"));
-				bean.setPrice(rs.getInt("PRICE"));
+				bean.setPrice(rs.getDouble("PRICE"));
 				bean.setQuantity(rs.getInt("QUANTITY"));
 				bean.setImage(rs.getString("IMAGE"));
 				products.add(bean);
